requirements.txt
google-genai>=1.0.0
opencv-python>=4.8.0
numpy>=1.24.0
pydantic>=2.0.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
torch>=2.0.0
timm>=0.9.0
